requirements.txt
# Epstein Network Analysis Project - Python Dependencies

# Image Processing (for cropping scripts)
Pillow>=10.0.0

# Progress bars for scripts
tqdm>=4.65.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file support

# JSON handling with better formatting
jsonlines>=3.1.0

# Future dependencies (for upcoming phases)
# Uncomment as needed:

# Neo4j Graph Database
# neo4j>=5.0.0
# py2neo>=2021.2.3

# Natural Language Processing
# spacy>=3.5.0
# transformers>=4.30.0

# Graph Analysis
# networkx>=3.0
# pyvis>=0.3.0

# API Development (for GraphRAG interface)
# fastapi>=0.100.0
# uvicorn>=0.23.0
# pydantic>=2.0.0

# Data Validation
# jsonschema>=4.17.0
# python-dateutil>=2.8.2

# Utilities
# python-dotenv>=1.0.0
# rich>=13.0.0  # For better terminal output

# Development Tools (optional)
# pytest>=7.3.0
# black>=23.0.0
# pylint>=2.17.0
# mypy>=1.3.0